Move matchers to separate file
