+++ frontend/src/metabase-lib/v1/expressions/recursive-parser.js
@@ -1,14 +1,9 @@
 import { t } from "ttag";
 
+import { isCaseOrIfOperator, isOptionsObject } from "./matchers";
 import { OPERATOR as OP, TOKEN, tokenize } from "./tokenizer";
 
-import {
-  MBQL_CLAUSES,
-  getMBQLName,
-  isCaseOrIfOperator,
-  isOptionsObject,
-  unescapeString,
-} from "./index";
+import { MBQL_CLAUSES, getMBQLName, unescapeString } from "./index";
 
 const COMPARISON_OPS = [
   OP.Equal,
